Fix all warnings in the AppleWin proj NB. 10 warnings still in libyaml proj (but there's no upstream update)

+++ source/Configuration/PropertySheetHelper.cpp
@@ -215,10 +215,10 @@ int CPropertySheetHelper::SaveStateSelectImage(HWND hWindow, const char* pszTitl
 			const char szAWS_EXT1[] = ".aws";
 			const char szAWS_EXT2[] = ".yaml";
 			const char szAWS_EXT3[] = ".aws.yaml";
-			const UINT uStrLenFile  = strlen(&szFilename[ofn.nFileOffset]);
-			const UINT uStrLenExt1  = strlen(szAWS_EXT1);
-			const UINT uStrLenExt2  = strlen(szAWS_EXT2);
-			const UINT uStrLenExt3  = strlen(szAWS_EXT3);
+			const size_t uStrLenFile  = strlen(&szFilename[ofn.nFileOffset]);
+			const size_t uStrLenExt1  = strlen(szAWS_EXT1);
+			const size_t uStrLenExt2  = strlen(szAWS_EXT2);
+			const size_t uStrLenExt3  = strlen(szAWS_EXT3);
 			if (uStrLenFile <= uStrLenExt1)
 			{
 				strcpy(&szFilename[ofn.nFileOffset+uStrLenFile], szAWS_EXT3);					// "file" += ".aws.yaml"
